fix: use exact service name match searching container labels
`findContainerByServiceName` incorrectly matched containers based on
service and router names when there were two services with similar
names, specifically when one service name was a prefix of the other.

Add a dot to the service and router needles and compare the labels with
HasPrefix() to ensure that only the exact matches get picked up.